recipe_app/requirements.txt
Django==3.1
django-filter==2.3.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.18.5
pandas==1.1.0
requests==2.24.0
scikit-learn==0.23.2
scipy==1.4.1
tensorflow==2.3.0
urllib3==1.25.10
whitenoise==5.2.0
djangorestframework~=3.11.1
gunicorn==20.0.4
